Main Algo cleanup
